Agregado los graphicos por rubro

+++ budget/views.py
@@ -57,9 +57,8 @@ def graphs_resume(request,pky=datetime.datetime.now().year,pkm=datetime.datetime
         for i in range(1,12):
             b = Budget.objects.filter(date__year=pky,date__month=i,category=cat)
             o = Outcome.objects.filter(date__year=pky,date__month=i,category=cat).aggregate(Sum('amount'))['amount__sum']
-            budgets.append(float(b[0].amount)) if b.count() > 0 else budgets.append(0)
-            outcomes.append(float(o)) if o else outcomes.append(0)
+            budgets.append(int(b[0].amount)) if b.count() > 0 else budgets.append(0)
+            outcomes.append(int(o)) if o else outcomes.append(0)
         tables.append([cat,budgets,outcomes])
     print(tables)
-        
     return render(request, 'graphics.html', {'color':'blue', 'today':today,'thismonth':thismonth,'thisyear':pky, 'tables':tables,'months':months})
